correction of the data type

+++ nehe/linux/lesson6/lesson6.c
@@ -23,12 +23,12 @@ int window;
 float xrot, yrot, zrot;
 
 /* storage for one texture  */
-int texture[1];
+GLuint texture[1];
 
 /* Image type - contains height, width, and data */
 struct Image {
-    unsigned long sizeX;
-    unsigned long sizeY;
+    unsigned int sizeX;
+    unsigned int sizeY;
     char *data;
 };
 typedef struct Image Image;
@@ -58,14 +58,14 @@ int ImageLoad(char *filename, Image *image) {
 	printf("Error reading width from %s.\n", filename);
 	return 0;
     }
-    printf("Width of %s: %lu\n", filename, image->sizeX);
+    printf("Width of %s: %u\n", filename, image->sizeX);
     
     // read the height 
     if ((i = fread(&image->sizeY, 4, 1, file)) != 1) {
 	printf("Error reading height from %s.\n", filename);
 	return 0;
     }
-    printf("Height of %s: %lu\n", filename, image->sizeY);
+    printf("Height of %s: %u\n", filename, image->sizeY);
     
     // calculate the size (assuming 24 bits or 3 bytes per pixel).
     size = image->sizeX * image->sizeY * 3;
